Adding imperial v3 rotator parts
Attaching 4 parts that are resized for imperial PVC sizes. I am
just merging the .stl files, leaving the changes to source out of
this as we do not want to manually keep up with parallel tracks
of source files as the desire is to make them configurable in the
future.  If the source for these imperial changes is needed, it
can be found at
https://github.com/cshields/satnogs-hardware/tree/imperial/src